## context.py * made credentials an optional argument when instantiating a consent * decrypting will now raise a ValueError if no credentials are available * putting to synapse was not recording the consent datetime but the dt for the update
## xtractor.py
* added kwargs to the TakeOutExtractor constructor to allow for
additional arguments, ie. the archive path
* added an additional class attribute to tell methods whether or not
we're working with the local filesystem or we need to download from
the participants google drive
* added private function to name files. one place to determine whether
we're working on a wsgi server or a local computer so name accordingly.
If on wsgi we use the wsgi tmp directory but if on a local we use the
current working directory
* added a print statement to the logging function to keep the user
updated with progress if running from local machine. won't print if
running on wsgi server
* added new function to load the archive from the local filesystem
rather than download from google
* added main method to make the file work as a standalone extractor or
to be called from other scripts